[OSC-1290] - rename mcd to dpo (apologies for extra readme corrections)
