+++ dpo/Shared.py
