+++ dpo/__main__.py
@@ -1,0 +1,3 @@
+from dpo.ModelChangeDetector import ModelChangeDetector
+
+ModelChangeDetector()
