+++ dpo/commands/InitialiseExecutionCommand.py
@@ -1,4 +1,4 @@
-from mcd.commands.BaseCommand import BaseCommand
+from dpo.commands.BaseCommand import BaseCommand
 
 
 class InitialiseExecutionCommand(BaseCommand):
