--- mcd/__main__.py
@@ -1,3 +1,0 @@
-from mcd.ModelChangeDetector import ModelChangeDetector
-
-ModelChangeDetector()
