+++ setup.py
@@ -1,6 +1,6 @@
 from setuptools import setup, find_packages
 
-setup(name='mcd',
+setup(name='dpo',
       version='0.0.2',
       packages=find_packages(),
       install_requires=[
